refactor : make useResolveSaveFunction
