+++ src/components/modals/CreateArticleModal/CreateArticleModal.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import styled from '@emotion/styled';
 
-import { useSetRecoilState, useRecoilValue } from 'recoil';
+import { useSetRecoilState } from 'recoil';
 import { openCreateModalStates } from 'src/recoil/permit';
-import { writeStates } from 'src/recoil/article';
 
 import Modal from '../Modal/Modal';
 import ImageUpload from 'src/components/ImageUpload/ImageUpload';
@@ -11,21 +10,14 @@ import Button from 'src/components/ui/Button/Button';
 import DescInput from 'src/components/Inputs/DescInput/DescInput';
 import TagInput from 'src/components/Inputs/TagInput/TagInput';
 
-import { useSaveArticle } from 'src/hooks/useCreatArticle';
-import { createArticle, updateArticle } from 'src/apis';
+import { useResolveSaveFunction } from 'src/hooks/useCreatArticle';
 
 const CreateArticleModal = () => {
   const setClose = useSetRecoilState(openCreateModalStates);
-  const writeState = useRecoilValue(writeStates);
-  const saveArticle = useSaveArticle();
+  const handleSaveArticle = useResolveSaveFunction();
 
   const handleModalClose = () => {
     setClose(false);
-  };
-
-  const handleSaveArticle = () => {
-    if (writeState == 'create') saveArticle(createArticle);
-    if (writeState == 'update') saveArticle(updateArticle);
   };
 
   return (
