+++ src/hooks/useCreatArticle.ts
@@ -2,10 +2,12 @@ import React from 'react';
 import { useRecoilState, useRecoilValue, useResetRecoilState } from 'recoil';
 import { useRouter } from 'next/router';
 
-import { articleStates } from 'src/recoil/article';
+import { articleStates, writeStates } from 'src/recoil/article';
 import { openCreateModalStates } from 'src/recoil/permit';
 
 import { SaveArticleFunction } from 'src/types/article';
+
+import { createArticle, updateArticle } from 'src/apis';
 
 export const useTitle = () => {
   const [articleElements, setArticleElements] = useRecoilState(articleStates);
@@ -60,9 +62,7 @@ export const useSaveArticle = () => {
   const resetModalState = useResetRecoilState(openCreateModalStates);
   const router = useRouter();
 
-  const handleSaveArticle = async (
-    saveArticleFunction: SaveArticleFunction,
-  ) => {
+  const saveArticle = async (saveArticleFunction: SaveArticleFunction) => {
     const res = await saveArticleFunction(articleElements);
 
     if (res.status == 200) {
@@ -70,6 +70,18 @@ export const useSaveArticle = () => {
       resetArticle();
       resetModalState();
     }
+  };
+
+  return saveArticle;
+};
+
+export const useResolveSaveFunction = () => {
+  const saveArticle = useSaveArticle();
+  const writeState = useRecoilValue(writeStates);
+
+  const handleSaveArticle = () => {
+    if (writeState == 'create') saveArticle(createArticle);
+    if (writeState == 'update') saveArticle(updateArticle);
   };
 
   return handleSaveArticle;
